+++ src/pages/index.tsx
@@ -4,26 +4,11 @@ import Layout from '@theme/Layout';
 import Link from '@docusaurus/Link';
 import useDocusaurusContext from '@docusaurus/useDocusaurusContext';
 import styles from './index.module.css';
+
+import HomepageHeader from '../components/HomepageHeader';
 import HomepageFeatures from '../components/HomepageFeatures';
 
-function HomepageHeader() {
-  const {siteConfig} = useDocusaurusContext();
-  return (
-    <header className={clsx('hero hero--primary', styles.heroBanner)}>
-      <div className="container">
-        <h1 className="hero__title">{siteConfig.title}</h1>
-        <p className="hero__subtitle">{siteConfig.tagline}</p>
-        <div className={styles.buttons}>
-          <Link
-            className="button button--secondary button--lg"
-            to="/docs/index">
-            Get Started
-          </Link>
-        </div>
-      </div>
-    </header>
-  );
-}
+
 
 export default function Home(): JSX.Element {
   const {siteConfig} = useDocusaurusContext();
